% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexInternal_functions.R
\name{BalancedGrps}
\alias{BalancedGrps}
\title{Function for resampling groups to equal sample size}
\usage{
BalancedGrps(GroupMembership, GroupSize = NA)
}
\arguments{
\item{GroupMembership}{a character or factor vector in the same order as the distance data to denote group membership.}

\item{GroupSize}{the sample size you wish the function to resample the data to. If NA supplied then the smallest sample size of the groups supplied will be used.}
}
\value{
Returns a list of 2 vectors. The first vector is the indexed location of the new equally size sample
}
\description{
This function takes a character of factor vector and returns a new vector of equal sample size
along with a vector of index positions to sort whatever other data might be associated with the
factor vector into the new equal sample size order. The user can also provide a user defined minimum
sample size if the user wishes the resampling procedure to sample to a smaller number of specimens
than the smallest sample size. In this way the function can also be used in a leave-p-out cross-
validation exercise.
}
\author{
Ardern Hulme-Beaman
}
