% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_descriptions.R
\docType{data}
\name{BirdsongData}
\alias{BirdsongData}
\title{Blue chaffinch birdsong dataset}
\format{
A list of 2 objects:
\describe{
  \item{SongDistMat}{A square matrix of dynamic time warping dissimilarities calculated among the songs of 116 individuals.}
  \item{Groups}{A dataframe of 4 columns including specimen IDs, Lat Long coordinates and Regional grouping.}
}
}
\usage{
BirdsongData
}
\description{
A birdsong dissimilarity dataset calculated using dynamic time warping. The
dataset includes 116 specimens from across Tenerife. These represent at least
three regional birdsong groups, with a fourth dispersed group from across the
centre of the island. The fourth group may not constitute a real group.
}
\keyword{datasets}
