% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_descriptions.R
\docType{data}
\name{DummyData}
\alias{DummyData}
\title{Dummy Data dataset}
\format{
A list of 3 objects:
\describe{
  \item{Coords}{A matrix of dummy coordinates}
  \item{Groups}{A vector of dummy group classifications}
  \item{DistMat}{A distance matrix of Euclidean distances among dummy specimens}
}
}
\usage{
DummyData
}
\description{
A dummy dataset for visualisation of difference between weighted and unweighted
k-NN analyses
}
\keyword{datasets}
