% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNNCrossValidation_functions.R
\name{KnnDistCV}
\alias{KnnDistCV}
\title{K-Nearest Neighbour correct cross-validation with distance input}
\usage{
KnnDistCV(
  DistMat,
  GroupMembership,
  K,
  Equal = TRUE,
  EqualIter = 100,
  SampleSize = NA,
  TieBreaker = c("Random", "Remove", "Report"),
  Weighted = FALSE,
  Verbose = FALSE,
  IgnorePrompts = FALSE
)
}
\arguments{
\item{DistMat}{is a square matrix of pairwise distances among all reference specimens.}

\item{GroupMembership}{a character or factor vector in the same order as the distance data to denote group membership.}

\item{K}{is the number of nearest neighbours that the method will use for assigning group classification.}

\item{Equal}{indicates where groups should be sampled to equal sample size}

\item{EqualIter}{sets the number of iterations resampling to equal sample size will be carried out.}

\item{SampleSize}{is the sample number that groups will be subsampled to if \code{Equal} is set to TRUE. The default is set to NA and will therefore use the smallest sample size of the groups provided.}

\item{TieBreaker}{is the method used to break ties if there is no majority resulting from K. Three methods are available('Random', 'Remove' and 'Report'): Random randomly returns one of tied classifications; Remove returns 'UnIDed' for the classification; Report returns a the multiple classifications as a single character string with tied classifications separated by '_'. NOTE: for correct cross-validation proceedures the results of both Report will be considered an incorrect identification even if one of the multiple reported classifications is correct.}

\item{Weighted}{is a logical TRUE or FALSE determining whether classification will be based on a weighted K. In this method a simple weighting system is used where weights are 1/distance, where distances is the distance of the considered neighbour from the unknown.}

\item{Verbose}{determines whether the cross-validation results for each reference specimen is returned. Note that if this is set to TRUE and Equal is set to TRUE the funtion will return a list with the results of each iteration which will slow the process dramatically and take a lot of local memory.}

\item{IgnorePrompts}{if both Verbose and Equal are set to TRUE, then the funciton will ask if you are sure you wish to continue; setting IgnorePrompts to TRUE will ignore this question.}
}
\value{
Returns a matrix of the leave-one-out classifications for all the specimens along with their known classificaiton.
}
\description{
This function NEEDS COMPLETION
}
\section{Citations}{


Ian L. Dryden (2016). shapes: Statistical Shape Analysis. R package version 1.1-13.
https://CRAN.R-project.org/package=shapes
}

\author{
Ardern Hulme-Beaman
}
\keyword{Geometric}
\keyword{distances}
\keyword{morphometric}
\keyword{shape}
