% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNNCrossValidation_functions.R
\name{KnnDistCVStepwise}
\alias{KnnDistCVStepwise}
\title{Stepwise K-Nearest Neighbour correct cross-validation with distance input}
\usage{
KnnDistCVStepwise(
  DistMat,
  GroupMembership,
  Kmax,
  Equal = TRUE,
  EqualIter = 100,
  SampleSize = NA,
  TieBreaker = c("Random", "Remove", "Report"),
  Verbose = FALSE,
  PrintProg = TRUE,
  PlotResults = TRUE
)
}
\arguments{
\item{DistMat}{is a square matrix of pairwise distances among all reference specimens.}

\item{GroupMembership}{a character or factor vector in the same order as the distance data to denote group membership.}

\item{Kmax}{This sets the maximum K that K will increase to stepwise.}

\item{Equal}{indicates where groups should be sampled to equal sample size}

\item{EqualIter}{sets the number of iterations resampling to equal sample size will be carried out.}

\item{SampleSize}{is the sample number that groups will be subsampled to if \code{Equal} is set to TRUE. The default is set to NA and will therefore use the smallest sample size of the groups provided.}

\item{TieBreaker}{is the method used to break ties if there is no majority resulting from K. Three methods are available('Random', 'Remove' and 'Report'): Random randomly returns one of tied classifications; Remove returns 'UnIDed' for the classification; Report returns a the multiple classifications as a single character string with tied classifications separated by '_'. NOTE: for correct cross-validation proceedures the results of both Report will be considered an incorrect identification even if one of the multiple reported classifications is correct.}

\item{Verbose}{determines whether the cross-validation results for each reference specimen is returned. Note that if this is set to TRUE and Equal is set to TRUE the funtion will return a list with the results of each iteration which will slow the process dramatically and take a lot of local memory.}

\item{PrintProg}{Only used when resampling to equal sample size is used (i.e. \code{Equal = TRUE}) and shows what iteration the function is on. Set to FALSE to ignore.}

\item{PlotResults}{logical when set to TRUE the results are plotted. When \code{Equal = TRUE} a polygon is plotted marking the 5th and 9th percentile.}
}
\value{
Returns a matrix of the leave-one-out classifications for all the specimens along with their known classification for both weighted and unweighted approaches.
}
\description{
This function NEEDS COMPLETION
}
\details{
When the \code{PrintProg} is set to TRUE, the \code{\link[svMisc]{progress}} function of the \code{svMisc} package is used.
}
\section{Citations}{


Ian L. Dryden (2016). shapes: Statistical Shape Analysis. R package version 1.1-13.
https://CRAN.R-project.org/package=shapes
}

\author{
Ardern Hulme-Beaman
}
\keyword{Geometric}
\keyword{distances}
\keyword{morphometric}
\keyword{shape}
