% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenIDing_functions.R
\name{KnnDistIDingBal}
\alias{KnnDistIDingBal}
\title{K-Nearest Neighbour multiple specimen identification with resampling to equal sample size}
\usage{
KnnDistIDingBal(
  DistMat,
  GroupMembership,
  UnknownIdentifier = "Unknown",
  SpecimenIDs,
  K,
  EqualIter = 100,
  TieBreaker,
  SampleSize = NA
)
}
\arguments{
\item{DistMat}{is a square matrix of pairwise distances among all reference specimens.}

\item{GroupMembership}{a character or factor vector in the same order as the distance data to denote group membership.}

\item{UnknownIdentifier}{the name used in the \code{GroupMembership} argument to denote specimens to be identified. Only one name can be supplied as Unknown; default is set to 'Unknown'.}

\item{SpecimenIDs}{should be a list of the specimens unique identifiers for all specimens in the \code{DistMat} object and ensuring they are in the same order as the \code{DistMat} object.}

\item{K}{is the number of nearest neighbours that the method will use for assigning group classification.}

\item{EqualIter}{sets the number of iterations resampling to equal sample size will be carried out.}

\item{TieBreaker}{is the method used to break ties if there is no majority resulting from K. Three methods are available('Random', 'Remove' and 'Report'): Random randomly returns one of tied classifications; Remove returns 'UnIDed' for the classification; Report returns a the multiple classifications as a single character string with tied classifications separated by '_'. NOTE: for correct cross-validation proceedures the results of both Report will be considered an incorrect identification even if one of the multiple reported classifications is correct.}

\item{SampleSize}{is the sample number that groups will be subsampled to if \code{Equal} is set to TRUE. The default is set to NA and will therefore use the smallest sample size of the groups provided.}
}
\value{
Returns a matrix of the leave-one-out classifications for all the specimens along with their known classificaiton.
}
\description{
This function is for a balanced KNN identification design applied to multiple unknown specimens.
Groups are resampled iteratively to equal sample size; note this is done by downsampleing the groups
to a sample size set by the user or if left to default to the sample size of the smallest groups.
Bootstrap resampling is not provided as an option because this can be problematic for nearest neighbour
approaches because of duplication of neighbours.
}
\author{
Ardern Hulme-Beaman
}
