% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenIDing_functions.R
\name{KnnDistIDingGroup}
\alias{KnnDistIDingGroup}
\title{k-Nearest Neighbour multiple specimen identification}
\usage{
KnnDistIDingGroup(
  DistMat,
  GroupMembership,
  UnknownIdentifier = "Unknown",
  K,
  TieBreaker
)
}
\arguments{
\item{DistMat}{is a square matrix of pairwise distances among all reference specimens.}

\item{GroupMembership}{a character or factor vector in the same order as the distance data to denote group membership.}

\item{UnknownIdentifier}{the name used in the \code{GroupMembership} argument to denote specimens to be identified. Only one name can be supplied as Unknown; default is set to 'Unknown'.}

\item{K}{is the number of nearest neighbours that the method will use for assigning group classification.}

\item{TieBreaker}{is the method used to break ties if there is no majority resulting from K. Three methods are available('Random', 'Remove' and 'Report'): Random randomly returns one of tied classifications; Remove returns 'UnIDed' for the classification; Report returns a the multiple classifications as a single character string with tied classifications separated by '_'. NOTE: for correct cross-validation proceedures the results of both Report will be considered an incorrect identification even if one of the multiple reported classifications is correct.}
}
\value{
Returns a matrix of the leave-one-out classifications for all the specimens along with their known classificaiton.
}
\description{
This function is for an unbalanced kNN identification design applied to multiple unknown specimens.
}
\author{
Ardern Hulme-Beaman
}
