% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecimenIDing_functions.R
\name{KnnIDingSingleInd}
\alias{KnnIDingSingleInd}
\title{K-Nearest Neighbour single specimen identification}
\usage{
KnnIDingSingleInd(
  X,
  K,
  GroupMembership,
  TieBreaker = c("Random", "Remove", "Report")
)
}
\arguments{
\item{X}{a numeric vector of distances from the unknown specimen to the all the reference specimens.}

\item{K}{is the number of nearest neighbours that the method will use for assigning group classification.}

\item{GroupMembership}{a character or factor vector in the same order as the distance data to denote group membership.}

\item{TieBreaker}{is the method used to break ties if there is no majority resulting from K. Three methods are available('Random', 'Remove' and 'Report'): Random randomly returns one of tied classifications; Remove returns 'UnIDed' for the classification; Report returns a the multiple classifications as a single character string with tied classifications separated by '_'. NOTE: for correct cross-validation proceedures the results of both Report will be considered an incorrect identification even if one of the multiple reported classifications is correct.}
}
\value{
Returns a matrix of the leave-one-out classifications for all the specimens along with their known classificaiton.
}
\description{
This function description NEEDS COMPLETION
}
\author{
Ardern Hulme-Beaman
}
