% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexInternal_functions.R
\name{KVote}
\alias{KVote}
\title{Function for calculating the majority vote}
\usage{
KVote(X, K, Weighting = FALSE, TieBreaker = c("Random", "Remove", "Report"))
}
\arguments{
\item{X}{a vector or matrix of the nearest neighbours. If \code{Weighting=TRUE} then a matrix is expected with the first column being the membership names of nearest neighbours and the second column should be the distances to each neighbour; if \code{Weighting=FALSE} then the just a vector of the membership names of nearest neighbours is required.}

\item{K}{is the number of nearest neighbours that the method will use for assigning group classification.}

\item{Weighting}{is a logical TRUE or FALSE determining whether classification will be based on a weighted K. In this method a simple weighting system is used where weights are 1/distance, where distances is the distance of the considered neighbour from the unknown.}

\item{TieBreaker}{is the method used to break ties if there is no majority resulting from K. Three methods are available('Random', 'Remove' and 'Report'): Random randomly returns one of tied classifications; Remove returns 'UnIDed' for the classification; Report returns a the multiple classifications as a single character string with tied classifications separated by '_'. NOTE: for correct cross-validation proceedures the results of both Report will be considered an incorrect identification even if one of the multiple reported classifications is correct.}
}
\value{
Returns the classification of the unknown individual to be identified.
}
\description{
This function carried out k-NN analyses on the provided data.
}
\author{
Ardern Hulme-Beaman
}
