% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{PlotStepwise}
\alias{PlotStepwise}
\title{Plotting function for stepwise results}
\usage{
PlotStepwise(
  StepwiseResultsMat,
  Probs = c(0.05, 0.95),
  PlotCol = "darkblue",
  Add = FALSE,
  Xlabel = "K"
)
}
\arguments{
\item{StepwiseResultsMat}{a}

\item{Probs}{a vector of two values denoting the upper and lower limit of the resampling to be plotted. Values must be between 0 and 1. Default is .05 and .95.}
}
\value{

}
\description{
This function plots the results of a stepwise analyses with resampling to equal sample size.
The function will plot the line connecting the means of each stepwise incriment increase of K
and will also plot the range around the mean (default upper and lower limit is set to the 5th
and 95th percentiles).
}
\author{
Ardern Hulme-Beaman
}
