% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapProcDist_functions.R
\name{ProcDistanceTable}
\alias{ProcDistanceTable}
\title{Wrapper for procdist function to output a distance table}
\usage{
ProcDistanceTable(
  RefShapes,
  LMDim = 2,
  ShapeArray = TRUE,
  Method = c("full", "partial", "Riemannian", "sizeandshape"),
  RefIDs = NA
)
}
\arguments{
\item{RefShapes}{either an array where rows are landmarks and columns are dimensions (i.e. x, y, and z) and slices are specimens or a matrix of coordinated where rows are specimens and columns are landmark coordinates in the format X1, Y1, X2, Y2... etc.}

\item{LMDim}{set to either 2 or 3 to denote whether the data is 2D or 3D. Default is set to 2.}

\item{ShapeArray}{either TRUE to denote that \code{TargetShape} is in matrix format and \code{Refshape} is an array, or FALSE to denote that \code{TargetShape} is a vector and \code{RefShape} is a matrix. Please do not use a combination of formats as the function does not handle this.}

\item{Method}{this is passed to the \code{\link[shapes]{procdist}} function of the \code{shapes} package and must be set to one of the following options: "full" for full Procrustes distance, "partial" for partial Procrustes distance, "Riemannian" for Riemannian shape distance, or "sizeandshape" for size-and-shape Riemannian/Procrustes distance.}

\item{RefIDs}{is a vector of the unique identifiers for each of the reference specimens in the reference dataset. These values will be used for naming the columns and rows. Default is set to NA and if a vector is not supplied the columns and rows will remain numbered consecutively.}
}
\value{
This function returns a square matrix of Procrustes distances, which is required for both the \code{IDbyDistanceRawDataCCV} and the \code{BoundaryFinder} functions.
}
\description{
This function builds a square matrix of pairwise procrustes distances among specimens using
the \code{\link[shapes]{procdist}} function from the \code{shapes} package.
}
\section{Citations}{


Ian L. Dryden (2016). shapes: Statistical Shape Analysis. R package version 1.1-13.
https://CRAN.R-project.org/package=shapes
}

\examples{
#RatDistMat <- ProcDistanceTable(Rpraetor$LMs)
}
\seealso{
\code{\link{ProcDistanceTablePar}}
}
