% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapProcDist_functions.R
\name{ShapeDist2Specimen}
\alias{ShapeDist2Specimen}
\title{Returns vector of Procrustes distances}
\usage{
ShapeDist2Specimen(
  RefShapes,
  TargetShape,
  LMDim = 2,
  ShapeArray = TRUE,
  Method = c("full", "partial", "Riemannian", "sizeandshape")
)
}
\arguments{
\item{RefShapes}{either an array where rows are landmarks and columns are dimensions (i.e. x, y, and z) and slices are specimens or a matrix of coordinated where rows are specimens and columns are landmark coordinates in the format X1, Y1, X2, Y2... etc.}

\item{TargetShape}{either a matrix where rows are landmarks and columns are dimensions (i.e. x, y, and z) or a vector of coordinated in the format X1, Y1, X2, Y2... etc.}

\item{LMDim}{set to either 2 or 3 to denote whether the data is 2D or 3D. Default is set to 2.}

\item{ShapeArray}{either TRUE to denote that \code{TargetShape} is in matrix format and \code{Refshape} is an array, or FALSE to denote that \code{TargetShape} is a vector and \code{RefShape} is a matrix. Please do not use a combination of formats as the function does not handle this.}

\item{Method}{this is passed to the \code{\link[shapes]{procdist}} function of the \code{shapes} package and must be set to one of the following options: "full" for full Procrustes distance, "partial" for partial Procrustes distance, "Riemannian" for Riemannian shape distance, or "sizeandshape" for size-and-shape Riemannian/Procrustes distance.}
}
\value{
The distance between the point of interest \code{TargetShape} to all other points in the reference material \code{RefShapes}.
}
\description{
This function is a wrapper for the \code{\link[shapes]{procdist}} function from the \code{shapes} package.
It constructs distances between a target shape and reference shapes using the method specified.
}
\section{Citations}{


Ian L. Dryden (2016). shapes: Statistical Shape Analysis. R package version 1.1-13.
https://CRAN.R-project.org/package=shapes
}

\author{
Ardern Hulme-Beaman
}
\keyword{Geometric}
\keyword{distances}
\keyword{internal}
\keyword{morphometric}
\keyword{shape}
