% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexInternal_functions.R
\name{TieBreakerFunction}
\alias{TieBreakerFunction}
\title{Function for breaking tied votes}
\usage{
TieBreakerFunction(X, TieBreaker)
}
\arguments{
\item{X}{a vector of the tied classifications to be considered.}

\item{TieBreaker}{is the method used to break ties if there is no majority resulting from K. Three methods are available('Random', 'Remove' and 'Report'): Random randomly returns one of tied classifications; Remove returns 'UnIDed' for the classification; Report returns a the multiple classifications as a single character string with tied classifications separated by '_'. NOTE: for correct cross-validation proceedures the results of both Report will be considered an incorrect identification even if one of the multiple reported classifications is correct.}
}
\value{
Returns SOMETHING
}
\description{
This function description NEEDS COMPLETION
}
\author{
Ardern Hulme-Beaman
}
