% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexInternal_functions.R
\name{BalancedGrps}
\alias{BalancedGrps}
\title{Function for resampling groups to equal sample size}
\usage{
BalancedGrps(GroupMembership, GroupSize = NA)
}
\arguments{
\item{GroupMembership}{a character or factor vector in the same order as the distance data to denote group membership.}

\item{GroupSize}{the sample size you wish the function to resample the data to. If NA supplied then the smallest sample size of the groups supplied will be used.}
}
\value{
Returns SOMETHING
}
\description{
This function description NEEDS COMPLETION
}
\author{
Ardern Hulme-Beaman
}
