% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{PlotStepwise}
\alias{PlotStepwise}
\title{Plotting function for stepwise results}
\usage{
PlotStepwise(
  StepwiseResultsMat,
  Probs = c(0.05, 0.5, 0.95),
  PlotCol = "darkblue",
  Add = FALSE,
  Xlabel = "K"
)
}
\value{

}
\description{
This function NEEDS COMPLETION
}
\author{
Ardern Hulme-Beaman
}
